appengine: Added handler "panel"
The Javascript now also uses this single URL path.
Rather than having a different path for each comic, accepting a single
argument for the link, there is now a single path which accepts two
arguments, one for the link and one for the comic.
The old handlers are still there, to allow for some backwards compatibility
while the Javascript update is pushed out