Added checkbox option for Google Reader
reader script will first check if it is running as the Chrome extension
by checking for the existence of the checkOption funciton. If so, then
it will be called just as in the other Javascript files for individual
sites. If the function doesn't exist, then it is assumed the script is
running in Greasemonkey, and so it will always run