reader: Fix for new location of ASP secret panels
Moved the JS code for ASP into common.js so that it can be used
in the files for both the ASP website and Google Reader. This
required a small change to the regex code, since the RSS feed
provides a slightly different image link format than the webpage itself.

Removed the appengine code for ASP, since it is no longer needed